+++ src/arguments/CoreCategoryChannel.ts
@@ -15,11 +15,6 @@ export class CoreArgument extends ExtendedArgument<'guildChannel', CategoryChann
 	public handle(channel: GuildChannel, context: ExtendedArgumentContext): ArgumentResult<CategoryChannel> {
 		return isCategoryChannel(channel)
 			? this.ok(channel)
-			: this.error({
-					parameter: context.parameter,
-					identifier: 'ArgumentCategoryChannelInvalidChannel',
-					message: 'The argument did not resolve to a category channel.',
-					context
-			  });
+			: this.error({ parameter: context.parameter, message: 'The argument did not resolve to a category channel.', context });
 	}
 }