+++ src/arguments/CoreNewsChannel.ts
@@ -15,11 +15,6 @@ export class CoreArgument extends ExtendedArgument<'guildChannel', NewsChannel> 
 	public handle(channel: GuildChannel, context: ExtendedArgumentContext): ArgumentResult<NewsChannel> {
 		return isNewsChannel(channel)
 			? this.ok(channel)
-			: this.error({
-					parameter: context.parameter,
-					identifier: 'ArgumentNewsChannelInvalidChannel',
-					message: 'The argument did not resolve to a news channel.',
-					context
-			  });
+			: this.error({ parameter: context.parameter, message: 'The argument did not resolve to a news channel.', context });
 	}
 }