+++ src/preconditions/DMOnly.ts
@@ -1,8 +1,11 @@
 import type { Message } from 'discord.js';
+import { Identifiers } from '../lib/errors/Identifiers';
 import { Precondition, PreconditionResult } from '../lib/structures/Precondition';
 
 export class CorePrecondition extends Precondition {
 	public run(message: Message): PreconditionResult {
-		return message.guild === null ? this.ok() : this.error({ message: 'You cannot run this command outside DMs.' });
+		return message.guild === null
+			? this.ok()
+			: this.error({ identifier: Identifiers.PreconditionDMOnly, message: 'You cannot run this command outside DMs.' });
 	}
 }